Load data from CSV
Note that the CSV has been omitted from the repo as I'm not convinced
(but also not *not* convinced) that we want the data living in there
(versus the internet, I guess? though it might still be nice to have a
committed copy of sample data to develop against).

Which is all to say that you'll need a copy of that csv in bin/data, and
it should be named covid_data.csv. Also note that there's really no
validation to speak of, so if we end up with poorly formatted data at
some point, it seems like that could fail more quietly than we'd like.